Tag version 2.15.2

+++ src/nmath/gamma.c
@@ -136,7 +136,7 @@ double gammafn(double x)
 	 * Reduce the interval and find gamma(1 + y) for 0 <= y < 1
 	 * first of all. */
 
-	n = (int) x;
+	n = x;
 	if(x < 0) --n;
 	y = x - n;/* n = floor(x)  ==>	y in [ 0, 1 ) */
 	--n;
